+++ app/app/javascript/adapters/PinwheelIncomeDataAdapter.ts
